requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
requests==2.18.4
solrpy3==0.98
urllib3==1.22
